Refactor : 파일 분리 및 코드 일부 수정 1. compontents/Home css 파일 분리 2. components/Schedule/Detail 'no-unde(import react -> import React)' 수정 3. App.css 주석처리 4. page 내 component 내용 삭제
