+++ src/App.css
@@ -1,8 +1,8 @@
-.App{
-  width: 100%;
-  height: 100vh;
-  max-width: 400px;
-  margin-left: auto;
-  margin-right: auto; 
-  background-color: #94bfff;
-}
+/* .App {
+    width: 100%;
+    height: 100vh;
+    max-width: 400px;
+    margin-left: auto;
+    margin-right: auto;
+    background-color: #94bfff;
+} */
