+++ src/pages/ContactInfo.js
@@ -4,6 +4,8 @@ import ContactInfoHeader from "../components/ContactInfo/ContactInfoHeader";
 import ContactInfoList from "../components/ContactInfo/ContactInfoList";
 
 const ContactInfo = () => {
+    const [selectedCategory, setSelectedCategory] = useState("안양캠퍼스");
+    
     const handleSelectCategory = (category) => {
         setSelectedCategory(category);
     };
