+++ src/pages/Faq.js
@@ -1,11 +1,7 @@
-import React from 'react';
+import React from "react";
 
 const Faq = () => {
-    return(
-        <div>
-        <h1>자주찾는질문</h1>
-       </div>
-    )
-}
+    return <></>;
+};
 
 export default Faq;
