+++ src/pages/Lost.js
@@ -1,11 +1,7 @@
-import React from 'react';
+import React from "react";
 
 const Lost = () => {
-    return(
-        <div>
-        <h1>캠퍼스 안내</h1>
-       </div>
-    )
-}
+    return <></>;
+};
 
 export default Lost;
