+++ src/styles/Home/Slider.css
